* libbase/utility.h: std::logf => std::log.

+++ libbase/utility.h
@@ -93,7 +93,8 @@ inline float flerp(float a, float b, float f) { return (b - a) * f + a; }
 
 //This is from C99.
 const float LN_2 = 0.693147180559945f;
-inline float	log2(float f) { return std::logf(f) / LN_2; }
+// the overridden log(f) will use logf IFF f is a float
+inline float	log2(float f) { return std::log(f) / LN_2; }
 inline int	fchop( float f ) { return (int) f; }	// replace w/ inline asm if desired
 inline int	frnd(float f) { return fchop(f + 0.5f); }	// replace with inline asm if desired
